top level functions instead of unnecessary objects
