+++ src/actions/Register.kt
@@ -5,13 +5,12 @@ import com.twoilya.lonelyboardgamer.InfoMissingException
 import com.twoilya.lonelyboardgamer.ServerResponse
 import com.twoilya.lonelyboardgamer.actions.commands.register.AddUser
 import com.twoilya.lonelyboardgamer.auth.JwtConfig
-import com.twoilya.lonelyboardgamer.auth.LoggedInService
+import com.twoilya.lonelyboardgamer.auth.isExist
 import com.twoilya.lonelyboardgamer.vk.VKConnector
-import io.ktor.application.call
-import io.ktor.request.receiveParameters
-import io.ktor.response.respond
-import io.ktor.routing.Routing
-import io.ktor.routing.post
+import io.ktor.application.*
+import io.ktor.request.*
+import io.ktor.response.*
+import io.ktor.routing.*
 
 fun Routing.registerRoute() {
     post("/register") {
@@ -20,7 +19,7 @@ fun Routing.registerRoute() {
         val vkAccessToken = parameters["VKAccessToken"] ?: throw InfoMissingException("No token provided")
         val userId = VKConnector.checkToken(vkAccessToken)
 
-        if (LoggedInService.isExist(userId)) {
+        if (isExist(userId)) {
             throw AuthorizationException("This user already exist")
         }
         AddUser.execute(userId, parameters)
