+++ src/actions/commands/profile/LogOut.kt
@@ -2,6 +2,7 @@ package com.twoilya.lonelyboardgamer.actions.commands.profile
 
 import actions.commands.TableCommand
 import com.twoilya.lonelyboardgamer.tables.UsersLoginInfo
+import com.twoilya.lonelyboardgamer.tables.dbQuery
 import org.jetbrains.exposed.sql.transactions.transaction
 import org.jetbrains.exposed.sql.update
 import org.joda.time.DateTime
